backend/requirements.txt
boto3
pypdf
PyPDF2
langchain
langchain-openai
chromadb
python-dotenv